+++ packages/ocaml-version.2.6.1/CHANGES.md
@@ -1,3 +1,13 @@
+## v2.6.1 (2020-07-15)
+
+* Fix `arch_of_string` to work with i386 strings (@avsm).
+
+## v2.6.0 (2020-07-14)
+
+* Mark trunk as 4.12.0 and add the 4.11 as a development beta (@avsm).
+* Add i386 architecture (@dinosaure @avsm)
+* Add a `arch_is_32bit` to determine if wordsize is 32-bits. (@avsm)
+
 ## v2.5.0 (2020-03-22)
 
 * Add OCaml 4.09.1 and 4.10.0 releases (@avsm @kit-ty-kate).